chore(test-config): refactor test config
