+++ packages/platform-test/e2e/pages/disease/associatedTargetsActions.spec.ts
@@ -3,6 +3,10 @@ import { AotfActions } from "../../../POM/objects/widgets/AOTF/aotfActions";
 
 test.describe("Disease Page - AOTF Actions", () => {
   test.beforeEach(async ({ page, baseURL, testConfig }) => {
+    //if no disease id, skip all tests
+    if (!testConfig.disease.primary) {
+      test.skip();
+    }
     await page.goto(`${baseURL}/disease/${testConfig.disease.primary}/associations`);
   });
