+++ packages/platform-test/e2e/pages/disease/associatedTargetsEvidenceWidgets.spec.ts
@@ -5,6 +5,10 @@ import { AotfTable } from "../../../POM/objects/widgets/AOTF/aotfTable";
 
 test.describe("Disease Page - AOTF Evidence Widgets", () => {
   test.beforeEach(async ({ page, baseURL, testConfig }) => {
+    //if no disease id, skip all tests
+    if (!testConfig.disease.primary) {
+      test.skip();
+    }
     await page.goto(`${baseURL}/disease/${testConfig.disease.primary}/associations`);
   });
