+++ packages/platform-test/e2e/pages/disease/associatedTargetsTable.spec.ts
@@ -4,6 +4,10 @@ import { AotfTable } from "../../../POM/objects/widgets/AOTF/aotfTable";
 
 test.describe("Disease Page - AOTF Table", () => {
   test.beforeEach(async ({ page, baseURL, testConfig }) => {
+    //if no disease id, skip all tests
+    if (!testConfig.disease.primary) {
+      test.skip();
+    }
     await page.goto(`${baseURL}/disease/${testConfig.disease.primary}/associations`);
   });
